requirements.txt
altair==6.0.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
lazy_imports==1.2.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.17.0
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
pluggy_sdk==1.0.0.post53
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.48
stack-data==0.6.3
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
